:recycle: [Refactor]: liveList를 모두 livePreviewList로 이름 변경
