+++ apps/client/src/pages/Home/ui/HomePage.tsx
@@ -1,10 +1,10 @@
-import { Banner, LiveList } from '@/widgets';
+import { Banner, LivePreviewList } from '@/widgets';
 
 export function HomePage() {
   return (
     <div className="flex flex-col justify-start w-full min-h-[calc(100vh-74px)]">
       <Banner />
-      <LiveList />
+      <LivePreviewList />
     </div>
   );
 }
