--- apps/client/src/widgets/LiveList/index.ts
@@ -1,1 +1,0 @@
-export { LiveList } from './LiveList';
